requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.12
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0
jinja2==3.1.4
